requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.6.4
entrypoints==0.4
exceptiongroup==1.2.0
Faker==23.1.0
favicon==0.7.0
fonttools==4.48.1
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.41
google-ai-generativelanguage==0.4.0
google-api-core==2.17.0
google-auth==2.27.0
google-generativeai==0.3.2
google-search-results==2.4.2
googleapis-common-protos==1.62.0
grpcio==1.60.1
grpcio-status==1.60.1
htbuilder==0.6.2
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.1.6
langchain-community==0.0.19
langchain-core==0.1.22
langchain-experimental==0.0.50
langchain-google-genai==0.0.8
langchainhub==0.1.14
langsmith==0.0.87
lxml==5.1.0
Markdown==3.5.2
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.8.2
mdurl==0.1.2
more-itertools==10.2.0
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
packaging==23.2
pandas==2.2.0
pillow==10.2.0
pinecone-client==3.0.2
prometheus-client==0.19.0
proto-plus==1.23.0
protobuf==4.25.2
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.1
pydantic_core==2.16.2
pydeck==0.8.1b0
Pygments==2.17.2
pymdown-extensions==10.7
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
st-annotated-text==4.0.1
streamlit==1.31.0
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.0
streamlit-embedcode==0.1.2
streamlit-extras==0.4.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.2
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.1
types-requests==2.31.0.20240125
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.2.0
validators==0.22.0
yarl==1.9.4
zipp==3.17.0
